ListPage削除=>SingleListPageをListPageに名称変更
