+++ lib/home_screen.dart
@@ -1,5 +1,4 @@
 import 'package:compare_2way/views/list/list_page.dart';
-import 'package:compare_2way/views/list/single_list_page.dart';
 import 'package:compare_2way/views/setting/setting_page.dart';
 import 'package:compare_2way/views/tag/tag_page.dart';
 import 'package:flutter/material.dart';
@@ -13,7 +12,7 @@ class HomeScreen extends StatefulWidget {
 class _HomeScreenState extends State<HomeScreen> {
 
   final _pages = [
-    SingleListPage(),
+    ListPage(),
     TagPage(),
     SettingPage(),
   ];
